chore(backend): move flup_backend into subfolder + add solution
