+++ backend/FlupBackend.sln
@@ -1,0 +1,26 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "flup_backend", "flup_backend\flup_backend.csproj", "{43AB6197-82F7-4B0B-AB04-B2800855A113}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A4679946-5D79-4ACF-84FA-DDFD7D032232}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4679946-5D79-4ACF-84FA-DDFD7D032232}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4679946-5D79-4ACF-84FA-DDFD7D032232}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4679946-5D79-4ACF-84FA-DDFD7D032232}.Release|Any CPU.Build.0 = Release|Any CPU
+		{43AB6197-82F7-4B0B-AB04-B2800855A113}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{43AB6197-82F7-4B0B-AB04-B2800855A113}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{43AB6197-82F7-4B0B-AB04-B2800855A113}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{43AB6197-82F7-4B0B-AB04-B2800855A113}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
