--- backend/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "DefaultConnection": "Data Source=sqlite.db"
-  },
-  "AllowedHosts": "*"
-}
