--- backend/flup_backend.csproj
@@ -1,29 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <RootNamespace>flup_backend</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="AutoMapper" Version="12.0.1"/>
-        <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.10"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.10">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.10"/>
-        <PackageReference Include="Stenn.EntityFrameworkCore.Extensions.DependencyInjection" Version="7.9.50"/>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0"/>
-    </ItemGroup>
-
-
-    <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-        <Exec Command="dotnet tool restore"/>
-        <Exec Command="dotnet swagger tofile --output swagger/swagger.json $(OutputPath)\$(AssemblyName).dll v1 "/>
-    </Target>
-
-</Project>
