+++ backend/flup_backend/src/Program.cs
@@ -1,0 +1,40 @@
+using Microsoft.EntityFrameworkCore;
+using flup_backend.Data;
+using flup_backend.Resolvers;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+
+builder.Services.AddControllers();
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+builder.Services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
+builder.Services.AddDbContext<AppDbContext>(options =>
+    options.UseSqlite(builder.Configuration.GetConnectionString("DefaultConnection")));
+builder.Services.AddTransient<ExistingPersonResolver>();
+builder.Services.AddTransient<ExistingPersonsResolver>();
+
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI(options =>
+    {
+        options.SwaggerEndpoint("/swagger/v1/swagger.json", "v1");
+        options.RoutePrefix = string.Empty;
+    });
+}
+else
+{
+    app.UseHttpsRedirection();
+    app.UseAuthorization();
+}
+
+app.MapControllers();
+
+app.Run();
